guards en chats2.controlelr

+++ backend/src/chats2/chats2.controller.ts
@@ -1,4 +1,4 @@
-import { Body, Controller, Delete, Get, HttpException, HttpStatus, Logger, Param, Post, Put, Req, UseGuards } from '@nestjs/common';
+import { Body, Controller, Delete, Get, HttpException, HttpStatus, Logger, Param, Post, Put, Req, UnauthorizedException, UseGuards } from '@nestjs/common';
 import { Jwt2faAuthGuard } from 'src/auth/jwt-2fa-auth.guard';
 import { getAuthToken, validateInput } from 'src/utils/utils';
 import { ChatMembershipDto, ChatMsgDto, ChatRoomDto, JoinChatRoomDto } from './chats.dto';
@@ -7,6 +7,8 @@ import { Chats2Service } from './chats2.service';
 import * as Joi from 'joi'
 import { UsersService } from 'src/users/users.service';
 import { JwtAuthenticatedGuard } from 'src/auth/jwt-authenticated-guard';
+import { HttpStatusCode } from 'axios';
+import { JwtAdminGuard } from 'src/auth/jwt-admin-guard';
 
 @Controller('chats')
 export class Chats2Controller {
@@ -47,9 +49,13 @@ export class Chats2Controller {
     }
 
     // get private chat for users
-    @UseGuards(JwtAuthenticatedGuard) //todo: add specific logic to only allow if requester is admin or one of the users
+    @UseGuards(JwtAuthenticatedGuard)
     @Post('rooms/direct')
-    async getPrivateRoom(@Body() body: any): Promise<ChatRoom> {
+    async getPrivateRoom(@Body() body: any, @Req() req): Promise<ChatRoom> {
+        const token = getAuthToken(req)
+        if (!token.hasRightsOverUser(token, await this.usersService.findOneById(body.user1)) &&
+        !token.hasRightsOverUser(token, await this.usersService.findOneById(body.user2)))
+            throw new UnauthorizedException(`Requester ${token.userId} does not have rights to retrieve private room between ${body.user1} and ${body.user2}`)
         //TODO añadir validacion de ids
         console.log("getPrivateRoom")
         Logger.log({ body })
@@ -81,12 +87,12 @@ export class Chats2Controller {
     }
 
     // update chat password
-    @UseGuards(JwtAuthenticatedGuard) //todo: update para que solo lo puedan hacer admin/owner or this chat admin/owner
+    @UseGuards(JwtAuthenticatedGuard)
     @Post('rooms/:roomId/password')
     @UseGuards(Jwt2faAuthGuard)
-    async updatePassword(@Req() req: any, @Param('roomId') roomId: number, @Body() data: any) {
-        const user = req.user
-        return this.chatsService.updateChatRoomPassword(user, roomId, data.password)
+    async updatePassword(@Req() req: any, @Param('roomId') roomId: number, @Body() data: { password: string }) {
+        const token = getAuthToken(req)
+        return this.chatsService.updateChatRoomPassword(token.userId, roomId, data.password)
     }
 
     // join chat room
@@ -117,10 +123,15 @@ export class Chats2Controller {
         return this.chatsService.inviteUser(user, roomId, data)
     }
 
-    @UseGuards(JwtAuthenticatedGuard) //todo: solo admin/owner o admin/owner de la room
+    @UseGuards(JwtAdminGuard) //solo admin/owner de la web, owner del chat lo puede borrar abandonando el chat
     @Delete('rooms/:id')
     async deleteRoom(@Param('id') id: number) {
-        this.chatsService.deleteRoom(id)
+        try {
+            await this.chatsService.deleteRoom(id)
+        } catch (error) {
+            if (error instanceof HttpException) throw (error)
+            throw new HttpException("delete room failed", HttpStatusCode.InternalServerError);
+        }
     }
 
     // // leave chat room
@@ -148,11 +159,11 @@ export class Chats2Controller {
     }
 
     // get user memberships
-    @UseGuards(JwtAuthenticatedGuard) //todo: update para qeu si no es user/admin solo pueda hacer esto consigo mismo, también revisar el primer if, ahora mismo parece que un admin/owner saca todas las memberships independinetemente del usuario elegido
+    @UseGuards(JwtAuthenticatedGuard)
     @Get('memberships/user/:userId')
     async findUserMemberships(@Req() req, @Param('userId') userId: string): Promise<ChatMembership[]> {
         const token = getAuthToken(req)
-        if (token.role === 'ADMIN' || token.role === 'OWNER') {
+        if (token.userId === userId && (token.role === 'ADMIN' || token.role === 'OWNER')) {
             const rooms = await this.chatsService.findAllChatRooms()
 
             const memberships : ChatMembership[] = rooms.map(room => {
@@ -170,46 +181,65 @@ export class Chats2Controller {
 
             return memberships
         }
-
+        if (!token.hasRightsOverUser(token, await this.usersService.findOneById(userId)))
+            throw new UnauthorizedException(`Requester ${token.userId} does not have rights to retrieve user's ${userId} memberships`)
         return await (this.chatsService.findUserMemberships(userId))
     }
 
+    //todo: importante updatear para que vaya con id y no username porq el username se puede cambiar y da problemas
     @UseGuards(JwtAuthenticatedGuard) //todo: solo admin/owner y admin/owner de la room
     @Post('memberships/ban')
-    async banUser(@Body() input: { userName: string, roomId: number }) {
+    async banUser(@Body() input: { userName: string, roomId: number }, @Req() req) {
+        const token = getAuthToken(req)
         const user = await this.usersService.findOneByName(input.userName);
 
         if (!user) throw new HttpException("USER_NOT_FOUND", HttpStatus.NOT_FOUND)
 
-        this.chatsService.setIsBanned(user.id, input.roomId, true);
+        const requesterMembership = await this.chatsService.findMembershipByUserAndRoom(token.userId, input.roomId)
+        if (token.hasRightsOverUser(token, user) && !requesterMembership && (!requesterMembership.isAdmin || !requesterMembership.isOwner))
+            this.chatsService.setIsBanned(user.id, input.roomId, true);
+        else
+            throw new UnauthorizedException(`Requester ${token.userId} does not have rights to ban user ${input.userName} in room ${input.roomId}`)
     }
 
-    @UseGuards(JwtAuthenticatedGuard) //todo: solo admin/owner y admin/owner de la room
+    @UseGuards(JwtAuthenticatedGuard)
     @Post('memberships/allow')
-    async allowUser(@Body() input: { userName: string, roomId: number }) {
+    async allowUser(@Body() input: { userName: string, roomId: number }, @Req() req) {
+        const token = getAuthToken(req)
         const user = await this.usersService.findOneByName(input.userName);
 
         if (!user) throw new HttpException("USER_NOT_FOUND", HttpStatus.NOT_FOUND)
 
-        this.chatsService.setIsBanned(user.id, input.roomId, false);
+        const requesterMembership = await this.chatsService.findMembershipByUserAndRoom(token.userId, input.roomId)
+        if (token.hasRightsOverUser(token, user) && !requesterMembership && (!requesterMembership.isAdmin || !requesterMembership.isOwner))
+            this.chatsService.setIsBanned(user.id, input.roomId, false);
+        else
+            throw new UnauthorizedException(`Requester ${token.userId} does not have rights to allow user ${input.userName} in room ${input.roomId}`)
     }
     
-    @UseGuards(JwtAuthenticatedGuard) //todo: solo admin/owner y admin/owner de la room
+    @UseGuards(JwtAuthenticatedGuard)
     @Post('memberships/promote')
-    async promoteUser(@Body() input: { userName: string, roomId: number }) {
+    async promoteUser(@Body() input: { userName: string, roomId: number }, @Req() req) {
+        const token = getAuthToken(req)
         const user = await this.usersService.findOneByName(input.userName);
         if (!user) 
             throw new HttpException("USER_NOT_FOUND", HttpStatus.NOT_FOUND)
+        
         const membership = await this.chatsService.findMembershipByUserAndRoom(user.id, input.roomId)
         if (!membership) 
             throw new HttpException("MEMBERSHIP_NOT_FOUND", HttpStatus.NOT_FOUND)
 
-        this.chatsService.setIsAdmin(user.id, input.roomId, true);
+        const requesterMembership = await this.chatsService.findMembershipByUserAndRoom(token.userId, input.roomId)
+        if (token.hasRightsOverUser(token, user) && !requesterMembership && (!requesterMembership.isAdmin || !requesterMembership.isOwner))
+            this.chatsService.setIsAdmin(user.id, input.roomId, true);
+        else
+            throw new UnauthorizedException(`Requester ${token.userId} does not have rights to promote user ${input.userName} in room ${input.roomId}`)
     }
 
-    @UseGuards(JwtAuthenticatedGuard) //todo: solo admin/owner y admin/owner de la room
+    @UseGuards(JwtAuthenticatedGuard)
     @Post('memberships/demote')
-    async demoteUser(@Body() input: { userName: string, roomId: number }) {
+    async demoteUser(@Body() input: { userName: string, roomId: number }, @Req() req) {
+        const token = getAuthToken(req)
         const user = await this.usersService.findOneByName(input.userName);
         if (!user) 
             throw new HttpException("USER_NOT_FOUND", HttpStatus.NOT_FOUND)
@@ -217,13 +247,17 @@ export class Chats2Controller {
         if (!membership) 
             throw new HttpException("MEMBERSHIP_NOT_FOUND", HttpStatus.NOT_FOUND)
 
-        this.chatsService.setIsAdmin(user.id, input.roomId, false);
+        const requesterMembership = await this.chatsService.findMembershipByUserAndRoom(token.userId, input.roomId)
+        if (token.hasRightsOverUser(token, user) && !requesterMembership && (!requesterMembership.isAdmin || !requesterMembership.isOwner))
+            this.chatsService.setIsAdmin(user.id, input.roomId, false);
+        else
+            throw new UnauthorizedException(`Requester ${token.userId} does not have rights to demote user ${input.userName} in room ${input.roomId}`)
     }
 
     // update membership
-    @UseGuards(JwtAuthenticatedGuard) //todo: solo admin/owner y el mismo usuario
+    @UseGuards(JwtAuthenticatedGuard)
     @Put('memberships/:id')
-    async updateMembership(@Param('id') id: number, @Body() data: ChatMembershipDto) {
+    async updateMembership(@Param('id') id: number, @Body() data: ChatMembershipDto, @Req() req) {
         // validateInput(Joi.object({
         //     userId: Joi.string().guid().required(),
         //     isOwner: Joi.boolean(),
@@ -231,20 +265,33 @@ export class Chats2Controller {
         //     isBanend: Joi.boolean(),
         //     isMuted: Joi.boolean()
         // }), data);
-        return this.chatsService.updateMembership(id, data)
+        const token = getAuthToken(req)
+
+        try {
+            const membershipOwner = await this.chatsService.findMembershipOwner(id)
+            if (token.hasRightsOverUser(token, membershipOwner))
+                return this.chatsService.updateMembership(id, data)
+        } catch (error) {
+            if (error instanceof HttpException) throw (error)
+            throw new HttpException("update membership failed", HttpStatusCode.InternalServerError);
+        }
     }
 
     // delete membership
-    @UseGuards(JwtAuthenticatedGuard) //todo: admin/owners + affected user
+    @UseGuards(JwtAuthenticatedGuard)
     @Delete('memberships/:id')
-    async deleteMembership(@Param('id') id: number) {
+    async deleteMembership(@Param('id') id: number, @Req() req) {
         if (id < 0)  // admin gets memberships for all rooms with id = -1
             return null 
+        const token = getAuthToken(req)
+        const membershipOwner = await this.chatsService.findMembershipOwner(id)
+        if (!token.hasRightsOverUser(token, membershipOwner))
+            throw new UnauthorizedException(`requester ${token.userId} cant delete membership ${id}`)
         return this.chatsService.deleteMembership(id)
     }
 
     // get chat messages for room
-    @UseGuards(JwtAuthenticatedGuard) //revisar pero extra checks en la logica
+    @UseGuards(JwtAuthenticatedGuard)
     @Get('/messages/:roomId')
     async findRoomMessages(@Req() req, @Param('roomId') roomId: number): Promise<ChatMsgDto[]> {
         const token = getAuthToken(req, false)
